feat : Step-3. 조인을 추가한 좋아요순 정렬 구현 & 쿼리 결과를 DTO로 받아오기

+++ src/main/java/io/dongvelop/springbootquerydsltutorial/board/dto/response/BoardListResponse.java
@@ -7,7 +7,7 @@ public record BoardListResponse(
         Long id,
         String title,
         Long views,
-        int likeCount,
+        Long likeCount,
         LocalDateTime createdAt
 ) {
 }
